fix: ZhouXun subsystem Roslyn diagnostics batch 4a (25 files)
- Remove empty Unity messages (UNT0001): CrossLine, MeshGen, GameLog,
  PlayerShakeEffect, UITextureHandler, RSManager, RCGen, EnergySheildHandler,
  GlobalMatMgr, SphereNetHandler, PointMesh, TreeGlowMgr, UITreeGridDebugger,
  VCMatManager, VCMeshMgr, CreationUI, ParticleScaler, VCEGizmoSphere
- Delete unused private fields (CS0169): CameraHitEffect (m_TargetPos),
  CameraShootEffect (m_TargetPos), CameraThirdPerson (m_CharacterOriginalPos)
- CA2000 wrap BinaryReader with leaveOpen:true: LSubTerrIO (shared FileStream),
  CRC64 (caller-owned stream), VCIsoData (3 sites: Export w/w_zip writers,
  Import r/r_unzip readers, ExtractHeader reader)
- CS0675 cast short to ushort before bit-shifts: LeastBox.Hash

+++ Assets/Scripts/ZhouXun/Voxel Creation/Scripts/Misc/LeastBox.cs
@@ -13,7 +13,7 @@ public struct IntBox
 	
 	public long Hash
 	{
-		get { return (xMin) | (zMin << 10) | (yMin << 20) | (xMax << 32) | (zMax << 42) | (yMax << 52); }
+		get { return (long)((ushort)xMin) | ((long)((ushort)zMin) << 10) | ((long)((ushort)yMin) << 20) | ((long)((ushort)xMax) << 32) | ((long)((ushort)zMax) << 42) | ((long)((ushort)yMax) << 52); }
 		set
 		{
 			xMin = (short)(value & 0x3ff);
